Refactor code, add `GradientScrollView`
